🏏🛰 Checkpoint ./komutlar/sil.js:28725120/1814

+++ komutlar/sil.js
@@ -10,12 +10,15 @@ let basari = ayarlar.basariliemoji
 let basarisiz = ayarlar.basarisizemoji
 let channel = message.channel;
 let member = message.author
+
+if(!args[0]) return message.channel.send(`${basarisiz} ${message.author}, Bir sayı belirtmelisin!`)
   
 var i = 0;
 message.delete();
 channel.messages.fetch().then(x => {
-x.filter(a => a.author.id === member.id).map(a => a).slice(0, 100.forEach(s => {
-i ++s.delete();
+x.filter(a => a.author.id === member.id).map(a => a).slice(args[0]).forEach(s => {
+i++
+s.delete();
 if(i === x.filter(a => a.author.id === member.id).map(a => a).slice(0, 100).length) {
 return message.channel.send(`${basari} ${message.author}, **${i}** Mesaj siliniyor.`);
 }
@@ -33,4 +36,4 @@ exports.conf = {
 
 exports.help = {
   name: "sil"
-}; 
+};
